requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
passlib[bcrypt]
python-jose
pillow
numpy
pytest
requests
rembg
